Moving proxy integration tests into single file to maintain 1:1 pairing of test files with libs
